[mdns] Add v2 component
This change introduces a CFv2 MDNS component under core, replacing the
existing legacy component that is in the sys realm under appmgr. It
leaves existing capability routing intact, so MDNS protocols are still
offered to the sys realm, and will be pruned if possible in a follow-up
change.

The legacy component remains because it is used in a suite of
integration tests that depend on the netemul runner, which has not yet
been migrated to CFv2.

Bug: 90433
Test: fx test mdns-tests
      fx test mdns-netemul-tests
Change-Id: I73ff5249fda177fb8b07db11605e9c5fca12e91b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/642888
Reviewed-by: Amanda Tait <atait@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Dale Sather <dalesat@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Mike "mc" Comfoltey <comfoltey@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>