[thermal] Per-sensor loads in ThermalLoadDriver
This change updates the ThermalLoadDriver node to maintain per-sensor
separation for thermal load values.

Previous behavior in this node was:
- poll each temperature handler input node to obtain its thermal load
- take the maximum of all sensor thermal loads
- send an UpdateThermalLoad message with this max thermal load value

The new behavior of the node is:
- poll each temperature handler input node to obtain its thermal load
- for each sensor's changed thermal load value, send an
  UpdateThermalLoad message with the new thermal load and corresponding
  sensor path

Motivation:
In future changes, the ThermalLoadDriver will be connected to the new
ThermalStateHandler node. The ThermalStateHandler requires per-sensor
thermal load values to inform the thermal state of connected clients.
This new per-sensor thermal load support in ThermalLoadDriver enables
the desired behavior of the ThermalStateHandler.

Bug: fxbug.dev/90157
Test: fx test power-manager-tests
Change-Id: Id0e52bee1df3be53076bd7f6c0c395ead5f5d5db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/630211
Reviewed-by: Didi She <didis@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>