[audio][tools] sig_gen ramp-start gain display
In the CLI utility signal_generator, in some cases the initial
gain is unspecified. In those cases, don't blindly dereference
the optional when displaying the summary of the run.

Also adjust the ramp-duration flag to be specified in seconds,
as is the case for overall signal duration.

fx shell signal_generator --gain=-45   --dur=1.234567 --ramp  \
                          --end-gain=0 --ramp-dur=1.2

Change-Id: I78318d7f1683d5f5ede8d38d21708767db8325c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/627876
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Alper Gungormusler <alperg@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>