[src][thread-state] Replace VLA with vector
As of clang revision 98936aee, UBSan detects an initialization of a
VLA of size zero. The function stores all handles sent via a channel
write into a VLA, but this particular test sends zero handles. A
simple fix is to just use a vector instead, which should still allow
for the test to timeout on the call to zx_object_wait_many.

Change-Id: Ife0602088baeb9e987e0df975ea0f605e4a68c6c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645582
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>