[phys] Introduce HandoffEntropy
This class job is to be the place where entropy sources from phys
are consumed. This involves copying out bytes and shredding
the contents.

For the zbi item the contents are zeroed after consumption.
For the cmdline entropy the contents are '0' after consumption.

Added zbi-handoff-entropy-tests to verify this behavior.

This CL does NOT introduce behavior changes with respect to the
existing mechanisms. That said, it is still not plugged in,
and that is reserved for a following change.

At this moment we dont have the mechanisms for exception
testing in phys, but we manually verified the behavior
with invalid inputs(invalid cmdline value, trying to
take entropy pool without enough entropy, etc).

This is in addition to existing emulator tests, that
already verified boot behavior under these scenarios.

Test: fx run-zbi-test qemu-phys-unittests.multiboot
Change-Id: I1455d0100817828fab4f45fc69b2a3c379c5edf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/612417
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>