[flatland] Modify flatland_benchmarks_test component kill order.
fxb/91066 is a flake where a RootPresenter from a previous test is
apparently not cleaned up properly.  Try topologically rearranging
the order that the test kills components, so that the RootPresenter
isn't restarted, and thus doesn't muck up the current test.

Also add "a11y-manager" to the list of killed components, since it
seems likely to be the culprit that is causing "root_presenter" to
restart.

Bug: 91066
Change-Id: I3124c90633175ee8bb8c3183941cdf597f3db969
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644036
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Nathan Rogers <nathanrogers@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>