[ffxutil][botanist] Fix isolated ffx environment.
There was a bug where botanist didn't provide the current environment to
the ffx instance so it didn't carry over the tmp variables that were set
to a different directory than /tmp and got cleaned up at the end of the
task. Thus, when botanist called ffx, it would create files in /tmp
without cleaning them up. To avoid this, we can have the FFXInstance
creation methods always include the os.Environ() in its environment.
Another bug was that botanist didn't carry over the ASCENDD env var to
testrunner because it doesn't actually set the env var in the current
environment but only when running ffx. Thus, testrunner would try to use
the same daemon as botanist by using the same config file, but it
wouldn't have the ASCENDD env var already set in its environment.

Bug: 84869, b/217261781
Change-Id: Id351c51c77ecafabc7e1e9e931364760b6944f8a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645742
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>