[check-licenses] Support generating template info for license run.
This is essentially what Misha created late last year.

I think this is the best way to implement the summary support given the
current organization of the tool. I will look to improve things in the
coming weeks as I make changes to check-licenses.

Change-Id: I4d20132edd0df1b92d6722de7d4ef147a4739aa4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/620945
Reviewed-by: Sean Cuff <seancuff@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>