[storage] Allocate MountOptions in the arena
An empty MountOptions was being created without any backing storage.
Passing the '-c' flag to the mount tool tried to set the compression
algorithm in the MountOptions causing a panic because of the missing
backing storage. Passing an arena to the MountOptions constructor
allocates backing storage in the arena.

Fixed: 93587
Change-Id: I7bf88738498c9e871039d8c3d0fca99e276e022b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645985
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>