[kernel][object] Avoid unexpected on_zero_handles
This changes zx_object_get_info(ZX_INFO_JOB_PROCESSES) to not report
processes that are in the INITIAL state with a handle count of 0. This
prevents zx_object_get_child being used on them, which can lead to a
premature invocation of on_zero_handles. This is important as processes
in the initial state perform actual work in on_zero_handles.

In particular there's a race with operations such as zx_process_create
where, prior to the user handle being created, the dispatcher is in the
job tree and can be found by zx_object_get_child. Until the user handle
is created the handle count is zero, and so the handle created by
get_child can cause the count to go from 0->1 and then 1->0 on close.
Result is the process returned by zx_process_create can be killed, due
to on_zero_handles running, before it's even returned to the user.

Bug: 76417, 93331

Change-Id: I01e4e12c37d70e9bda28b1f5af1ecfb3b6841756
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/643402
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Carlos Pizano <cpu@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>