[reachability] Soft-migrate out of management
As part of CFv2 migration, reachability is moving out of
//src/connectivity/management and into //src/connectivity/network.
Subsequently, management will be nearly entirely vestigial and will
therefore be removed. Prepare for this change by introducing an empty
group target //src/connectivity/network/netcfg:config-default and adding
it to core.gni. This is step 1 of 5 of the soft-migration plan detailed
in the linked bug.

Bug: 89115

Change-Id: Ib63142171c45b382aedac2fa37cb4a7539f1c990
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646104
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Amanda Tait <atait@google.com>