[fidl][cpp] Fix async completer use-after-free
The async completer in the unified server bindings contains an object
that borrows `this`. This is problematic for move operations as the
`this` pointer will be invalidated.

The easiest fix is to not use member fields borrowing `this`.

Test: added fidl_cpp_gen_api_tests similar to the LLCPP counterpart

Change-Id: I72ed609dfa3ae765d85201d15552c19f58c091db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646051
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>