[rust][clippy] Add flag to not fail on deny lints
For tricium we'd like to be able to let the clippy builds pass even
when they contain deny lints. This will also be helpful down the line
when doing global clippy runs when enabling new deny lints so we don't
have to bother with telling the build to ignore errors to see them all.

Change-Id: I4a81690c6a34b4dbba827e5a91f5c62e7bcc7976
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645584
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>
Commit-Queue: Joseph Ryan <josephry@google.com>