[kernel][vm] Pausable VMAR enumerator
This commit refactors the callback based enumeration in
EnumerateChildInternalLocked into a stateful enumerator that yields
regions. Further, this stateful enumerator is extended to be able to
pause and resume and tolerate the vmar hierarchy changing. The pause
and resume functionality is not yet used beyond tests.

Bug: 84616

Change-Id: Iedda8e177d9c4a1db577ff387cd510aa10b43d5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/640902
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>