[ffx] Add config setting to enable the repository server
This unifies fx and fserve to use the ffx config setting
repository.server.mode to determine which repository server. This avoids
the risk that the ffx daemon and the fx/fserve wrappers are in sync with
which repository server to use.

In order to do this, it:

* The ffx repository daemon service starts the server if
  repository.server.mode is "ffx".
* fserve to use repository.server.mode instead of fserve.server-mode.
* fx switches from using the feature flag "legacy_serve" to using
  repository.server.mode.

Fixed: 93547

Change-Id: I9860864d3ff92876913c52ff554633c5371739b5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645684
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Reviewed-by: Misha Gridnev <gridman@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>