[tools][fuzz] Improve QEMU boot monitoring logic
When starting a new instance, the previous code relied on `ps` to
determine whether QEMU had terminated early. However, because undercoat
detaches from QEMU and does not wait on the subprocess, `ps` will still
report an exited QEMU as a defunct process. This resulted in a deadlock
condition when a boot timed out. Instead of attempting to work around
this, which is OS-specific, this CL refactors the boot monitoring logic
to make it unnecessary.

Another benefit of the change is that some testing-specific `ps`
proxying code, which in fact masked the presence of this error, has been
removed.

Change-Id: Ia018d90f38a09c65f3a2bf4e7334c0b98e162417
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644033
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Cameron Finucane <eep@google.com>