[ffx] Add a health check to ffx emu
This CL adds code to wait for the emulator to respond to requests from
the daemon before returning from the launcher. It times out after 60
seconds, configurable via the --startup-timeout flag or `ffx config set
emu.start.timeout`. As long as we're waiting, the launcher maintains
control of the stdin/stdout streams, so a user can respond to prompts
(such as "sudo" in the shortleash-upscript).

Bug: fxbug.dev/91977
Bug: fxbug.dev/93172
Change-Id: I15d347afffaf797160a8a07242a7c2ac70e5e04f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/644784
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>