[ffx][test] Add missing artifact type to schema
Adds a missing artifact type to the schema.
This also adds a macro to autogenerate a list of all the variants for
an enum containing no fields, and uses the list to test all variants
against the schema. The macro is loosely based on a 'decodable_enum'
macro defined in //src/connectivity. This should at least block
submission when a variant not recognized in the schemas is added.

Testing: new unit tests, manually verify that adding a new variant
without updating schemas breaks tests.

Bug: 81195

Change-Id: I8ac6dc32547d0315ca35d7c007cc151ef18b0aaa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646590
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Satsuki Ueno <satsukiu@google.com>