[ffx] Reuse image files for persisted emulators
The ffx emu shutdown command has a "--persist" flag, which leaves the
working directory in place after the emulator is terminated. These files
could be reused on future instances, but aren't currently. The "start"
command always re-stages the files, so the persisted versions are
overwritten.

This CL adds a check to the start-up code, such that if image files
already exist in the working directory, they aren't overwritten. The
upshot is that an emulator that was shutdown with the --persist flag can
effectively be "restarted", just by using the same instance name when
calling ffx emu start.

Bug: fxbug.dev/80063
Change-Id: I755246d5c4eeacd715f3a0bae697d6e04b04f86c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645569
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>