[build][python] Ability to specify the field serializer fn
ints with a value of 0 were not being serialized as part of the check
to see if a value was present or not.  This makes the logic detect
that situation, and adds the ability to specify a serializer fn with
a class decorator: '@serialize_field_as(my_field=str)'

Also adds the '@serialize_json' and '@serialize_dict' decorators that
I had previously written, but left in a stash.

Includes tests for all the various code-paths.

Change-Id: Ic75ae0d3bf05c335aa8ab06c8100a7cc826e7991
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/641711
Reviewed-by: Aidan Wolter <awolter@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>