[wlan][policy] refactor EMWA RSSI recording
Previously, the connected state in state_machine.rs would process
periodic signal reports and call into the SavedNetworksManager to record
the data and SavedNetworksManager would track the EWMA RSSI and SNR and
velocity. After this refactor the EWMA RSSI and SNR tracking happens
during the connected state, which works better for future changes and
also means that the smoothed values and velocities are based on values
from the same connections.

Test: added checks for the new behavior in the connected state
  fx test wlancfg-tests
Change-Id: I5597a0f13ff664d2fe5c7851ef230055a66097a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646657
Reviewed-by: Hayden Nix <haydennix@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>