[netcfg] Reject all unknown configurations
The `#[serde(deny_unknown_fields)]` attribute is set on the `Config`
type, but is not present on any of its members, though these are all
strict. This CL sets the attribute on all contained structs and enums
such that unrecognized options will be rejected anywhere in the
configuration.

It may also be useful to have runtime tests for this behavior in case
future changes forget to add this attribute. But that's outside the
scope of this CL.

Bug: 93845
Test: added new test validating all types contained within `Config`.
Change-Id: I7fa2473f37bf30eedb7ea1196161d755d72bc095
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646767
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Devon O'Dell <dhobsd@google.com>