[kernel][vm][oom] Trigger OOM when PMM runs out of pages
This change adds an optional feature (off by default) that will
trigger a system OOM when the PMM fails to allocate for a
non-borrowing request because it has run out of memory.

This change is a stopgap aimed at improving system resilience.

Because of the asynchronous, non-latching nature of memory pressure
levels, it's possible for the system to completely run of pages, fail
one or more allocations, free some memory, and get back to a normal
pressure level before the memory watchdog has had an opportunity to
take action (like commit to an OOM reboot).

Currently, it's not possible for a program to gracefully recover from
allocation failure during a page fault.  Even if it were, allocation
failure during a syscall is generally unexpected and likely to cause
failures that leave the system in a wedged state.

This change updates the PMM and MemoryWatchdog to effectively latch
when the PMM runs out of pages and ensures that the configured OOM
response will occur.

Added new test case to oom_tests.

Bug: 93282
Test: ./out/core.x64/host_x64/oom_tests
Change-Id: I203cbbd0144a734efd5af9f0ec99f0e6984f0894
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645109
Reviewed-by: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>