[cm] Introduce shutdown::Component trait.
Before this change, much of the shutdown logic gets its data about the
component that's shutting down from a `ComponentDecl`. This strategy
precludes using information like dynamic offers in computing the
shutdown order, and also makes it so that the tests for
`process_component_dependencies` can't know anything about dynamic
instances.

Behaviorally, this change is a no-op, but it replaces that use of
`ComponentDecl` with use of a new `shutdown::Component` trait. This
layer of indirection will allow us to account for dynamic offers and
children in subsequent CLs.

Bug: 84678

Change-Id: I4084bccd4dd38e05f88cdc38e47a91ea54cc5500
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/646702
Reviewed-by: Derek Gonyeo <dgonyeo@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>