[devshell] fx format-code should not fail on one unknown extension
A fairly recent change was made to fail if `fx format-code` is passed a
"single file" with an unknown extension, but for multiple files, it was
supposed to format the known extensions and fail silently on unknown
extensions.

This change didn't work as intended because the check for multiple files
assumed the `FILES` variable was an array.

Usually, it was not an array, so the number of elements was never
greater than 1, despite the number of files.

It turns out that `FILES` is set in multiple places, and most don't make
it an array. (One does.) And many other places that used `FILES` used it
assuming it was an array. This only worked (most of the time) because
the spaces were expanded into multiple words upon evaluation.

In general, I'm surprised more things weren't breaking.

I updated all uses (set and evaluation) to use the array syntax, and
updated the test to validate that, with this change, now a list of files
with some known and some unknown extensions will still format the known
ones.

Change-Id: I0e503201520824e51314629bd70fc171da16759f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647567
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Commit-Queue: Rich Kadel <richkadel@google.com>