[ffx][product] Remove workaround for bug 93799
The "start_up_args_template": "emulator_flags.json.template",
is now in the cb container. The workaround is not necessary
for current builds.

This workaround could be kept longer to support prior builds,
but there's not many, so it seems reasonable to remove this
workaround.

Bug: 93799
Change-Id: I520c9c2839aab13267724fedc3bf943d9a99d7be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647662
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>