[pwauth] Bump timeout in integration tests
On ASAN builds, we've observed this flake occur from time to time where
the devmgr never manages to bring up ramctl within 20 seconds.  We only
see this flake on the arm64-asan builder.

We have reason to believe that these integration tests are simply
running out of CPU time, due to the fact that many of them are doing the
work of full password verification, which is intentionally
computationally expensive to limit the rate at which attackers can
attempt to crack passwords.

To attempt to patch around the high flake rate, try increasing the
timeout, which should allow slightly more core time on slow ASAN
builders that are running VMs.  Perhaps in the fullness of time, we'll
move these tests to only run on real hardware, but we'll want adequate
CQ coverage if we go that route.

Bug: 90826
Bug: 93675
Change-Id: I0bb785f68fbd40365f7924c29ac7594b93a93956
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647347
Fuchsia-Auto-Submit: Drew Fisher <zarvox@google.com>
Reviewed-by: Ben Wright <benwright@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>