[ffx] Change default behavior for name parameters
The attached bug indicated that stop and show should default to the
only running emulator when there is only one running, rather than having
a fixed default value. This simplifies the workflow for named emulators,
even if the default name is not used at start.

This CL modifies the behavior of both stop and show, such that if no
name is provided, the command will call get_all_instances, then if
only one value is returned that value will be used for the command. If
no instances are found, a message will be printed to indicate this to
the user, and the command will terminate. If multiple emulators are
running, a message asking the user to utilize the name parameter is
printed, and the command terminates. We also add return codes for both
subcommands, so that test infra can differentiate between the various
failure modes.

Bug: fxbug.dev/93472
Change-Id: Ie807c9b784c57f705a9ffd6a2a89adcdf451566a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647384
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>