[tests][oom] Fix test race where we wait for a wrong message
This change fixes a bug in the oom_test that was introduced with the
kernel.oom.trigger-on-alloc-failure feature.  Depending on the exact
nature of the OOM event we may or may not observe the memory pressure
state change message.

Test: Ran host_x64/oom_tests 10 times.
Fixed: 94019
Change-Id: I2986959c97e0919dd03189f44813b3969c5248fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647801
Fuchsia-Auto-Submit: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>