[component_manager] fix DestroyChild race
The DestroyChild action was using non-instanced monikers. This means
that if a DestroyChild action was racing with a DestroyChild and a
CreateChild action, in rare cases the duplicate DestroyChild action
could destroy the instance created by the CreateChild instance.

To fix this, refactor the DestroyChild action to work with
InstancedChildMoniker, and have it check the instance of the child
returned by `get_live_child` to confirm that it's operating on the
correct child.

Generally actions that perform work on children should use instanced
monikers to reference those children, to remove
time-of-check-time-of-use issues where the live children set is changed
between when the action is created and when it's able to be
registered/run.

Fixed: 92354
MULTIPLY: network-test-realm-integration-tests
Change-Id: I0af778a47fb96d7e2b296af23a5d44638b05563a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645942
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Derek Gonyeo <dgonyeo@google.com>