[session] Host tmp memfs for chrome in ermine
Chrome now uses `tmp` storage defined in the `ermine` shell component,
which is hosted by a sibling component, `memfs`.

This ensures that the contents of `tmp` are cleared when memfs is
stopped. Ermine must be stopped recursively for this to happen
on logout.

Bug: 92782
Change-Id: Ib28b650cba208065fbb1a8283cfb6826ecfe798b
Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/647841
Fuchsia-Auto-Submit: Yegor Pomortsev <ypomortsev@google.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>