[carnelian] default to scenes for rendering
Add a new method to the ViewAssistant trait to return an optional
mutable reference to the view's scene. Components which wish to use
scenes can implement this method to remove some boilerplate code.

In the future this will be used to allow user input events to be
delegated to scenes without requiring the view assistant to do the
forwarding. By this methods things like drag and drop can be easily
implemented in the scene.

As part of this CL I moved the button, font_metrics, layout_gallery and
spinning_square examples to take advantage of the new trait method.
Other scene-using samples were not using layout, so could not be easily
migrated since the new trait method calls layout, which fails when
there's no root group. This area needs to be cleaned up a bit then
everything scene-using can be moved over.

Testing: manual and existing unit tests.

Change-Id: I73b2ca73263d8be59ef67b029c97f48219088248
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638968
Reviewed-by: David Reveman <reveman@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>