[power] Create SystemPowerModeHandler node
Creates a new node called SystemPowerModeHandler. This node is
responsible for serving three new protocols:

  fuchsia.power.clientlevel.Connector - allows a client service to
  connect a fuchsia.power.clientlevel.Watcher server end to receive
  power level changes of a desired fuchsia.power.clientlevel.ClientType

  fuchsia.power.systemmode.Requester - allows a component to request
  changes to the currently active system power modes

  fuchsia.power.systemmode.ClientConfigurator - allows a component to
  reconfigure the system power mode configuration for a client

The node reads in the system power mode configuration file during
initialization and uses this config to determine the appropriate power
level for a client based on the currently active system power modes.

This CL simply adds the code for the new node, but it is not
instantiated on any product (yet).

Bug: 91839
Test: fx test power-manager-tests
Change-Id: I2a57ccfab90d2e50fa3904a366f0f0ba469d8b15
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645685
Reviewed-by: Didi She <didis@google.com>
Reviewed-by: Onath Dillinger <claridge@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>