[ffx] Add support for integers to FfxConfigBacked
While setting up the timeout for ffx emu's liveness check, we found that
FfxConfigBacked only has two supported modes: String and Float. We used
the Float type to get our feature running, but integer types are really
a better fit.

This CL adds an IntegerType to the ConfigValueType struct, as well as
the necessary conversion routines, with a backing type of u64.

Change-Id: I7269829f4a06ce7ecea61fd3d429c2400caeb789
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647327
Reviewed-by: Matt Boetger <boetger@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>