[sherlock][power] Do not reconfigure VDDEE_800 rail on init.
As part of PWM initialization we reconfigure all PWMs including
PWMAO_B which controls the VDDEE_800 rail which is already
configured by the bootloader. We should be leaving this rail as-is.

Setting the `protect` flag on the PWM prevents our PWM driver from
reconfiguring this rail.

Test: Smoke test completed by booting on a device
umangshah@ also verified the rail using a scope.

Change-Id: Ib54fd0e0c8c0286102366a32b1826ead6e773cec
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648004
Reviewed-by: Eric Holland <hollande@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>