[audio_core] Simplify Mixer::Mix
Mixer::Mix currently returns a bool which is true iff the source buffer
was fully consumed. This bool is redundant:

  bool consumed = mixer.Mix(dest_buffer, dest_count, &dest_offset,
                            source_buffer, source_count, &source_offset,
                            accumulate);
  assert(consumed == (source_offset + pos_filter_width > source_count));

To prepare for a future change in which we stop tracking whether a
buffer is fully consumed, this CL removes Mixer::Mix's return value.

Bug: 50669
Bug: 73306
Change-Id: I8d3c202e9340b2c885566cdf0ada1742373a7d50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639699
Reviewed-by: Alper Gungormusler <alperg@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>