[kernel][vm] Require PageRequest
Enforces the requirement of GetPage that a PageRequest be passed in if
fault flags are set, removing the temporary carveout for anonymous
VMOs.

Also annotates these locations that now need to pass in pagerequests as
needing validation that locks are not held when those pageqrequests are
in use.

Bug: 94078

Change-Id: I332f75e3354f11f3fab631623b0e16c9c2a43e1d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647866
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>