[netstack3] Remove DeviceId from network sim
The `DummyNetwork` simulation in the `context::testutil` module has
`SendMeta` and `RecvMeta` parameters for the metadata produced when
sending and consumed when receiving, respectively. It also has a
`DeviceId` parameter. This latter parameter is redundant - when a
network is being simulated at the level of an entire stack complete with
devices, `SendMeta` represends the device ID of the outbound device on
the sending host, and `RecvMeta` represents the device ID of the inbound
device on the receiving host. (In other scenarios, such as simulating
communication between two transport-layer sockets, there is no notion of
a device). `DeviceId` is redundant, and so is removed in this commit.

Change-Id: If9eae174cdfdd963fe2da173e985cf87ec3c9405
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648029
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Joshua Liebow-Feeser <joshlf@google.com>