[mdns] move deprecated services to alternate impl class
This CL moves the implementation of discoverable services that we
plan to deprecate into their own implementation class. The services
moved are:

    fuchsia.net.mdns.Resolver
    fuchsia.net.mdns.Subscriber
    fuchsia.net.mdns.Publisher

These will all return under new names with additional features.
Because some of the new services will have the same method names
as the deprecated services, implementing them in the same class
is complicated. Also, this split will make it easier to remove the
deprecated service implementations once we are done with them.

Test: no behavior change, all tests pass

Bug: 93320

Change-Id: Ia167ddcc227586e96b56f49edc1d9eba42fd5c44
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/647943
Reviewed-by: Robert Quattlebaum <rquattle@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>