[kernel][lockup] Don't deref an invalid shadow call stack pointer
When generating a DAP backtrace for a locked up CPU, don't read the
memory pointed to by the SCSP unless it's a kernel address and mapped.

Pull the backtrace generation logic out of DumpRegistersAndBacktrace
and into its own method GetBacktraceFromDapState so that it can be
easily tested.

Add test for GetBacktraceFromDapState.

Bug: 93993
Test: k ut lockup_detector
Change-Id: I0b7c8be3802ee4ccfa7540cdd40df7855916e4b8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648093
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>