[f2fs] Revisit the fsync and checkpoint operations
This CL enables the asynchronous writeback in checkpoint and fsync.
At checkpoint time, all necessary Pages for consistency are flushed
to disk in order asynchronously, but the dirty Pages for files are
written out at fsync or during writeback or shutdown. This way
speeds up checkpointing as well as fsync by flushing as many as
necessary of dirty Pages. To reduce lock scope, it separates
the writeback logic from the FileCache traversal. It reduces
the time for f2fs-slow-fs-test by more than 50%.

Test: fx test f2fs-slow-fs-tests f2fs-fs-tests

Change-Id: Iaae1f70800c128ed4c8ceb2e74ef5b5e14a7a3a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/638641
Reviewed-by: Aaron Drew <ripper@google.com>
Commit-Queue: Kim Seokhwan <sukka.kim@samsung.com>