[x86][acpi] Move GetBti to FIDL.
We switch from allowing devices to specify an arbitrary PCI
bus/device/function to get an IOMMU for to determining one for
ourselves.

The new GetBti implementation adds some constraints over the old one:

* Callers cannot be on a bus (i.e. they must only be ACPI devices). We
  will relax this restriction in the future, most likely to only prevent
  PCI devices from getting an ACPI BTI.
* Callers will always get back a dummy BTI with a meaningless (but
  unique) BTI ID.

In the future we will add support for determining the IOMMU to be used
for an ACPI device based on the DMAR entries.

Bug: 80244
Change-Id: I3448348872da49435d18a69066019726f2bdba7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/639014
Reviewed-by: Alex Legg <alexlegg@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>