[debugger] Fix a bug in DebuggedProcess::OnResume
The IPC specifies that in ResumeRequest, an entry with a process koid
and a 0 thread koid will resume all threads of the given process.
However, this is not correctly implemented in DebuggedProcess::OnResume
and causes a problem when debug_agent detaches from a process.

Fixed: 90060
Change-Id: I96fb884cf215f0f2f94ffa283bc8904925bffd20
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645116
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Dangyi Liu <dangyi@google.com>